apps/apim-simulator/pyproject.toml
[project]
name = "apim-simulator"
version = "0.4.0"
description = "Config-driven Azure API Management simulator for local Docker-first testing"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "defusedxml>=0.7.1",
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
    "pyjwt[crypto]>=2.10.1",
    "uvicorn>=0.34.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "opentelemetry-exporter-otlp-proto-http>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-httpx>=0.61b0",
    "opentelemetry-instrumentation-logging>=0.61b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "coverage[toml]>=7.6.0",
    "diff-cover>=9.2.0",
    "ruff>=0.8.0",
]
mcp = [
    "mcp==1.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
exclude-newer = "7 days"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "contract(*ids): associates a test with one or more contract ids from contracts/contract_matrix.yml",
    "integration: requires external services (e.g. Keycloak) running",
]

[tool.coverage.run]
branch = true
source = ["app"]
relative_files = true

[tool.coverage.report]
fail_under = 75
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
